requirements.txt
sentence-transformers
numpy
spacy
flair
nltk
vaderSentiment
requests
beautifulsoup4
pyap
phonenumbers
email-validator
pytesseract
fitz
python-docx
pdf2image
plotly
seaborn
matplotlib
faiss-cpu
scikit-learn
boto3
selenium
webdriver-manager
fuzzywuzzy
flask
